ai_engine/requirements.txt
pandas
prophet
pymongo
python-dotenv
openpyxl